open gov refactoring - treasury overview component
